requirements.txt
trame
trame-vuetify
trame-plotly
lume-model
pre-commit
pandas
numpy